requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM & Embeddings
openai==1.10.0
anthropic==0.18.1

# RAG & Vector Store
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16
chromadb==0.4.22

# Document Processing
pymupdf==1.23.21
pypdf==4.0.1
pillow==10.2.0

# Text Processing
sentence-transformers==2.3.1
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1

# Monitoring & Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0